work on examples and types
